Move `StacksTransactionSigner` back to `stackslib`
