+++ stackslib/src/chainstate/stacks/transaction.rs
@@ -14,6 +14,177 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+use crate::chainstate::stacks::{
+    Error, StacksPrivateKey, StacksPublicKey, StacksTransaction, StacksTransactionSigner,
+    TransactionAuth, TransactionAuthField, TransactionSpendingCondition,
+};
+use crate::net::Error as net_error;
+
+/// Pop the last auth field
+fn pop_auth_field(
+    condition: &mut TransactionSpendingCondition,
+) -> Option<TransactionAuthField> {
+    match condition {
+        TransactionSpendingCondition::Multisig(ref mut cond) => cond.pop_auth_field(),
+        TransactionSpendingCondition::OrderIndependentMultisig(ref mut cond) => {
+            cond.pop_auth_field()
+        }
+        TransactionSpendingCondition::Singlesig(ref mut cond) => cond.pop_signature(),
+    }
+}
+
+impl StacksTransactionSigner {
+    pub fn new(tx: &StacksTransaction) -> StacksTransactionSigner {
+        StacksTransactionSigner {
+            tx: tx.clone(),
+            sighash: tx.sign_begin(),
+            origin_done: false,
+            check_oversign: true,
+            check_overlap: true,
+        }
+    }
+
+    pub fn new_sponsor(
+        tx: &StacksTransaction,
+        spending_condition: TransactionSpendingCondition,
+    ) -> Result<StacksTransactionSigner, Error> {
+        if !tx.auth.is_sponsored() {
+            return Err(Error::IncompatibleSpendingConditionError);
+        }
+        let mut new_tx = tx.clone();
+        new_tx.auth.set_sponsor(spending_condition)?;
+        let origin_sighash = new_tx.verify_origin()?;
+
+        Ok(StacksTransactionSigner {
+            tx: new_tx,
+            sighash: origin_sighash,
+            origin_done: true,
+            check_oversign: true,
+            check_overlap: true,
+        })
+    }
+
+    pub fn resume(&mut self, tx: &StacksTransaction) {
+        self.tx = tx.clone()
+    }
+
+    pub fn disable_checks(&mut self) {
+        self.check_oversign = false;
+        self.check_overlap = false;
+    }
+
+    pub fn sign_origin(&mut self, privk: &StacksPrivateKey) -> Result<(), net_error> {
+        if self.check_overlap && self.origin_done {
+            // can't sign another origin private key since we started signing sponsors
+            return Err(net_error::SigningError(
+                "Cannot sign origin after sponsor key".to_string(),
+            ));
+        }
+
+        match self.tx.auth {
+            TransactionAuth::Standard(ref origin_condition) => {
+                if self.check_oversign
+                    && origin_condition.num_signatures() >= origin_condition.signatures_required()
+                {
+                    return Err(net_error::SigningError(
+                        "Origin would have too many signatures".to_string(),
+                    ));
+                }
+            }
+            TransactionAuth::Sponsored(ref origin_condition, _) => {
+                if self.check_oversign
+                    && origin_condition.num_signatures() >= origin_condition.signatures_required()
+                {
+                    return Err(net_error::SigningError(
+                        "Origin would have too many signatures".to_string(),
+                    ));
+                }
+            }
+        }
+
+        let next_sighash = self.tx.sign_next_origin(&self.sighash, privk)?;
+        self.sighash = next_sighash;
+        Ok(())
+    }
+
+    pub fn append_origin(&mut self, pubk: &StacksPublicKey) -> Result<(), net_error> {
+        if self.check_overlap && self.origin_done {
+            // can't append another origin key
+            return Err(net_error::SigningError(
+                "Cannot append public key to origin after sponsor key".to_string(),
+            ));
+        }
+
+        self.tx.append_next_origin(pubk).map_err(net_error::from)
+    }
+
+    pub fn sign_sponsor(&mut self, privk: &StacksPrivateKey) -> Result<(), net_error> {
+        if let TransactionAuth::Sponsored(_, ref sponsor_condition) = self.tx.auth {
+            if self.check_oversign
+                && sponsor_condition.num_signatures() >= sponsor_condition.signatures_required()
+            {
+                return Err(net_error::SigningError(
+                    "Sponsor would have too many signatures".to_string(),
+                ));
+            }
+        }
+
+        let next_sighash = self.tx.sign_next_sponsor(&self.sighash, privk)?;
+        self.sighash = next_sighash;
+        self.origin_done = true;
+        Ok(())
+    }
+
+    pub fn append_sponsor(&mut self, pubk: &StacksPublicKey) -> Result<(), net_error> {
+        self.tx.append_next_sponsor(pubk).map_err(net_error::from)
+    }
+
+    pub fn pop_origin_auth_field(&mut self) -> Option<TransactionAuthField> {
+        match self.tx.auth {
+            TransactionAuth::Standard(ref mut origin_condition) => {
+                pop_auth_field(origin_condition)
+            }
+            TransactionAuth::Sponsored(ref mut origin_condition, _) => {
+                pop_auth_field(origin_condition)
+            }
+        }
+    }
+
+    pub fn pop_sponsor_auth_field(&mut self) -> Option<TransactionAuthField> {
+        match self.tx.auth {
+            TransactionAuth::Sponsored(_, ref mut sponsor_condition) => {
+                pop_auth_field(sponsor_condition)
+            }
+            _ => None,
+        }
+    }
+
+    pub fn complete(&self) -> bool {
+        match self.tx.auth {
+            TransactionAuth::Standard(ref origin_condition) => {
+                origin_condition.num_signatures() >= origin_condition.signatures_required()
+            }
+            TransactionAuth::Sponsored(ref origin_condition, ref sponsored_condition) => {
+                origin_condition.num_signatures() >= origin_condition.signatures_required()
+                    && sponsored_condition.num_signatures()
+                        >= sponsored_condition.signatures_required()
+                    && (self.origin_done || !self.check_overlap)
+            }
+        }
+    }
+
+    pub fn get_tx_incomplete(&self) -> StacksTransaction {
+        self.tx.clone()
+    }
+
+    pub fn get_tx(&self) -> Option<StacksTransaction> {
+        if self.complete() {
+            Some(self.tx.clone())
+        } else {
+            None
+        }
+    }
+}
 
 #[cfg(test)]
 mod test {
